% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_xx.R
\name{assert_proportion_value}
\alias{assert_proportion_value}
\title{Check whether \code{x} is a proportion: number between 0 and 1.}
\usage{
assert_proportion_value(x, include_boundaries = FALSE)
}
\arguments{
\item{x}{(\code{any})\cr object to test.}

\item{include_boundaries}{(\code{flag})\cr whether to include boundaries when testing
for proportions.}
}
\description{
This function is copied from internal tern function assert_proportion_value, to avoid import
}
