% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{basic_pagdf}
\alias{basic_pagdf}
\title{Basic/spoof pagination info data frame}
\usage{
basic_pagdf(
  rnames,
  labs = rnames,
  rnums = seq_along(rnames),
  extents = 1L,
  rclass = "NA",
  parent_path = "root",
  fontspec = font_spec()
)
}
\arguments{
\item{rnames}{(\code{character})\cr vector of row names.}

\item{labs}{(\code{character})\cr vector of row labels. Defaults to \code{rnames}.}

\item{rnums}{(\code{integer})\cr vector of row numbers. Defaults to \code{seq_along(rnames)}.}

\item{extents}{(\code{integer})\cr number of lines each row requires to print. Defaults to 1 for all rows.}

\item{rclass}{(\code{character})\cr class(es) for the rows. Defaults to \code{"NA"}.}

<<<<<<< HEAD
\item{parent_path}{character. parent path that all rows should be "children of",
defaults to \code{"root"}, and generally should not matter to end users.}

\item{fontspec}{font_spec. A font_spec object
specifying the font information to use for
calculating string widths and heights, as
returned by \code{\link[=font_spec]{font_spec()}}}
=======
\item{parent_path}{(\code{string})\cr parent path that all rows should be "children of".
Defaults to \code{"root"}, and generally should not matter to end users.}
>>>>>>> main
}
\value{
A \code{data.frame} suitable for use in both the \code{MatrixPrintForm} constructor and the pagination machinery.
}
\description{
Returns a minimal pagination info \code{data.frame} (with no info on siblings, footnotes, etc.).
}
\examples{
basic_pagdf(c("hi", "there"))

}
