% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_xx.R
\name{count_fraction}
\alias{count_fraction}
\alias{format_sas_count_fraction}
\alias{format_count_fraction}
\alias{format_r_count_fraction}
\alias{format_sas_count_denom_fraction}
\alias{format_r_count_denom_fraction}
\title{Count_fraction}
\usage{
format_sas_count_fraction(x, output, d = 1)

format_r_count_fraction(x, output, d = 1)

format_sas_count_denom_fraction(x, output, d = 1)

format_r_count_denom_fraction(x, output, d = 1)
}
\arguments{
\item{x}{\code{numeric}\cr with elements \code{num} and \code{fraction} or \code{num}, \code{denom} and \code{fraction}.}

\item{output}{required for rtables interface}

\item{d}{numeric(1). Number of digits to round fraction to (default=1)}
}
\value{
A string in the format \verb{count / denom (ratio \%)}. If \code{count} is 0, the format is \code{0}. If fraction is >0.99, the format is \verb{count / denom (>99.9\%)}
}
\description{
A family of formatting functions for presenting a count together with fraction (and denominator) with special consideration when count is 0, or fraction is 1.
\cr \code{format_r_count_fraction} is a version that is similar to tern function tern::format_count_fraction_fixed_dp
}
\examples{
format_sas_count_fraction(c(7, 0.7))
format_sas_count_fraction(c(70000, 0.9999999))
format_sas_count_fraction(c(70000, 1))

format_sas_count_denom_fraction(c(3, 2000, 3 / 2000))
format_sas_count_denom_fraction(c(3, 2000, 3 / 2000))
format_r_count_denom_fraction(c(3, 2000, 3 / 2000))
format_r_count_denom_fraction(c(3, 2000, 3 / 2000))
}
