% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{default_horizontal_sep}
\alias{default_horizontal_sep}
\alias{default_hsep}
\alias{set_default_hsep}
\title{Default horizontal separator}
\usage{
default_hsep()

set_default_hsep(hsep_char)
}
\arguments{
\item{hsep_char}{(\code{character(1)})\cr character that will be set in the R environment
options as the default horizontal separator. Must be a single character. Use
\code{getOption("formatters_default_hsep")} to get its current value (\code{NULL} if not set).}
}
\value{
\code{unicode} 2014 (long dash for generating solid horizontal line) if in a
locale that uses a UTF character set, otherwise an ASCII hyphen with a
once-per-session warning.
}
\description{
The default horizontal separator character which can be displayed in the current
charset for use in rendering table-like objects.
}
\examples{
default_hsep()
set_default_hsep("o")
default_hsep()

}
