% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{default_hsep}
\alias{default_hsep}
\title{Default horizontal Separator}
\usage{
default_hsep()
}
\value{
\code{unicode} 2014 (long dash for generating solid horizontal line)
if in a locale that uses a UTF character set, otherwise an ASCII hyphen
with a once-per-session warning.
}
\description{
The default horizontal separator character which can be
displayed in the current \code{charset} for use in rendering table-likes.
}
\examples{
default_hsep()
}
