% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{default_rounding}
\alias{default_rounding}
\alias{set_default_rounding}
\title{Default rounding type}
\usage{
default_rounding()

set_default_rounding(round_type = c("iec", "sas"))
}
\arguments{
\item{round_type}{(\code{string})\cr single character value to set the default rounding type. It can be
either \code{"iec"} or \code{"sas"} for IEC 60559 or SAS rounding (nearest-value rounding), respectively.}
}
\value{
The default rounding type (\code{"iec"} if not set).
}
\description{
The default rounding type for numeric values in any formatting outputs like \code{\link[=format_value]{format_value()}}.
}
\examples{
default_rounding()
set_default_rounding("sas")
default_rounding()

}
