% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporters.R
\name{export_as_flextable}
\alias{export_as_flextable}
\title{Create a FlexTable object representing an rtables TableTree}
\usage{
export_as_flextable(
  tt,
  paginate = FALSE,
  lpp = NULL,
  cpp = NULL,
  ...,
  colwidths = propose_column_widths(matrix_form(tt, indent_rownames = TRUE)),
  tf_wrap = !is.null(cpp),
  max_width = cpp,
  total_width = 10
)
}
\arguments{
\item{tt}{TableTree (or related class). A TableTree object representing a populated table.}

\item{paginate}{logical(1). Should \code{tt} be paginated and exported as
multiple flextables. Defaults to \code{FALSE}}

\item{lpp}{numeric. Maximum lines per page including (re)printed header and context rows}

\item{cpp}{numeric(1). Number of columns per page}

\item{...}{Passed directly to \code{\link{paginate_table}}}

\item{colwidths}{numeric vector. Column widths for use with vertical pagination. Currently ignored.}

\item{total_width}{numeric(1). Total width in inches for the resulting
flextable(s). Defaults to 5.}
}
\value{
a flextable object
}
\description{
Create a FlexTable object representing an rtables TableTree
}
\examples{
library(dplyr)
library(rtables)

analysisfun <- function(x, ...) {
    in_rows(row1 = 5,
            row2 = c(1, 2),
            .row_footnotes = list(row1 = "row 1 - row footnote"),
            .cell_footnotes = list(row2 = "row 2 - cell footnote"))
}

lyt <- basic_table(title = "Title says Whaaaat", subtitles = "Oh, ok.",
               main_footer = "ha HA! Footer!") \%>\%
split_cols_by("ARM") \%>\%
analyze("AGE", afun = analysisfun)

tbl <-  build_table(lyt, ex_adsl)
ft <- export_as_flextable(tbl)
ft
}
