% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{fmt_config}
\alias{fmt_config}
\title{Format Configuration}
\usage{
fmt_config(format = NULL, na_str = "NA", align = "center")
}
\arguments{
\item{format}{character(1) or function. A format label (string) or \code{formatter} function.}

\item{na_str}{character(1). String that should be displayed in place of missing values.}

\item{align}{character(1). Alignment values should be rendered with.}
}
\value{
An object of class \code{fmt_config} which contains the following elements:
\itemize{
\item \code{format}
\item \code{na_str}
\item \code{align}
}
}
\description{
Format Configuration
}
\examples{
fmt_config(format = "xx.xx", na_str = "-", align = "left")
fmt_config(format = "xx.xx - xx.xx", align = "right")

}
