% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{horizontal_sep}
\alias{horizontal_sep}
\alias{default_hsep}
\alias{set_default_hsep}
\title{Default horizontal Separator}
\usage{
default_hsep()

set_default_hsep(hsep_chr)
}
\arguments{
\item{hsep_chr}{character(1). Character that will be set in the R environment
options as default for creating the horizontal separator. It needs to be
single character. Use \code{getOption("formatters_default_hsep")} to get its current
value (\code{NULL} if not set).}
}
\value{
\code{unicode} 2014 (long dash for generating solid horizontal line)
if in a locale that uses a UTF character set, otherwise an ASCII hyphen
with a once-per-session warning.
}
\description{
The default horizontal separator character which can be
displayed in the current \code{charset} for use in rendering table-likes.
}
\examples{
default_hsep()

}
