% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifnotlen0}
\alias{ifnotlen0}
\alias{\%||\%}
\title{\verb{\%||\%} If length-0 alternative operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{ANY. Element to select only if it is not length 0}

\item{b}{ANY. Element to select if \code{a} is length 0}
}
\value{
\code{a}, unless it  is length 0, in which case  \code{b} (even in the
case \code{b} is also length 0)
}
\description{
\verb{\%||\%} If length-0 alternative operator
}
\examples{
6 \%||\% 10

character() \%||\% "hi"

NULL \%||\% "hi"
}
