% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if a value is a whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{(\code{numeric(1)})\cr a numeric value.}

\item{tol}{(\code{numeric(1)})\cr a precision tolerance.}
}
\value{
\code{TRUE} if \code{x} is within \code{tol} of zero, \code{FALSE} otherwise.
}
\description{
Check if a value is a whole number
}
\examples{
is.wholenumber(5)
is.wholenumber(5.00000000000000001)
is.wholenumber(.5)

}
