% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jjcsformats.R
\name{jjcs_format_xx_fct}
\alias{jjcs_format_xx_fct}
\title{Title: Function factory for xx style formatting}
\usage{
jjcs_format_xx_fct(roundmethod = c("SAS", "R"))
}
\arguments{
\item{roundmethod}{(\code{string})\cr choice of rounding methods. Options are:
\itemize{
\item \code{SAS}: the underlying rounding method is \code{tidytlg::roundSAS}, where \cr
roundSAS comes from this Stack Overflow post https://stackoverflow.com/questions/12688717/round-up-from-5
\item \code{R}: the underlying rounding method is \code{round}
}}
}
\value{
\code{jjcs_format_xx_fct()} format function that can be used in rtables formatting calls
}
\description{
A function factory to generate formatting functions for value
formatting that support the xx style format and control the rounding method
}
\examples{
jjcsformat_xx_SAS <- jjcs_format_xx_fct(roundmethod="SAS")
jjcsformat_xx <- jjcsformat_xx_SAS
format_value(c(1.453),jjcsformat_xx("xx.xx"))
format_value(c(1.453,2.45638),jjcsformat_xx("xx.xx (xx.xxx)"))

}
\seealso{
Other JJCS formats: 
\code{\link{Count_fraction}}
}
\concept{JJCS formats}
