% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{list_formats}
\alias{list_formats}
\alias{list_valid_format_labels}
\alias{list_valid_aligns}
\title{List of currently supported formats and vertical alignments}
\usage{
list_valid_format_labels()

list_valid_aligns()
}
\value{
\itemize{
\item \code{list_valid_format_labels()} returns a nested list, with elements listing the supported 1d, 2d,
and 3d format strings.
}

\itemize{
\item \code{list_valid_aligns()} returns a character vector of valid vertical alignments.
}
}
\description{
We support \code{xx} style format labels grouped by 1d, 2d, and 3d.
Currently valid format labels cannot be added dynamically. Format functions
must be used for special cases.
}
\examples{
list_valid_format_labels()

list_valid_aligns()

}
