% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_valid_aligns}
\alias{list_valid_aligns}
\title{Utility to list valid cell alignments}
\usage{
list_valid_aligns()
}
\description{
Utility to list valid cell alignments
}
\examples{
list_valid_aligns()

}
