% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{list_valid_format_labels}
\alias{list_valid_format_labels}
\title{List with currently support 'xx' style format labels grouped by 1d, 2d and 3d}
\usage{
list_valid_format_labels()
}
\value{
A nested list, with elements listing the supported 1d, 2d, and 3d format strings.
}
\description{
Currently valid format labels can not be added dynamically. Format functions
must be used for special cases
}
\examples{

list_valid_format_labels()

}
