% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_form.R
\name{MatrixPrintForm-class}
\alias{MatrixPrintForm-class}
\title{Matrix Print Form - Intermediate Representation for ASCII Table Printing}
\description{
Matrix Print Form - Intermediate Representation for ASCII Table Printing
}
