% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_form.R
\name{mf_strings}
\alias{mf_strings}
\alias{mf_spans}
\alias{mf_aligns}
\alias{mf_display}
\alias{mf_formats}
\alias{mf_rinfo}
\alias{mf_cinfo}
\alias{mf_has_topleft}
\alias{mf_lgrouping}
\alias{mf_rfnotes}
\alias{mf_nlheader}
\alias{mf_nrheader}
\alias{mf_colgap}
\alias{mf_fontspec}
\alias{mf_fontspec<-}
\alias{mf_col_widths}
\alias{mf_col_widths<-}
\alias{mf_strings<-}
\alias{mf_spans<-}
\alias{mf_aligns<-}
\alias{mf_display<-}
\alias{mf_formats<-}
\alias{mf_rinfo<-}
\alias{mf_cinfo<-}
\alias{mf_lgrouping<-}
\alias{mf_rfnotes<-}
\alias{mf_nrheader<-}
\alias{mf_colgap<-}
\alias{mf_ncol}
\alias{mf_nrow}
\alias{mf_ncol<-}
\alias{ncol,MatrixPrintForm-method}
\alias{mpf_has_rlabels}
\alias{mf_has_rlabels}
\title{Getters and setters for aspects of \code{MatrixPrintForm} objects}
\usage{
mf_strings(mf)

mf_spans(mf)

mf_aligns(mf)

mf_display(mf)

mf_formats(mf)

mf_rinfo(mf)

mf_cinfo(mf)

mf_has_topleft(mf)

mf_lgrouping(mf)

mf_rfnotes(mf)

mf_nlheader(mf)

mf_nrheader(mf)

mf_colgap(mf)

mf_fontspec(mf)

mf_fontspec(mf) <- value

mf_col_widths(mf)

mf_col_widths(mf) <- value

mf_strings(mf) <- value

mf_spans(mf) <- value

mf_aligns(mf) <- value

mf_display(mf) <- value

mf_formats(mf) <- value

mf_rinfo(mf) <- value

mf_cinfo(mf) <- value

mf_lgrouping(mf) <- value

mf_rfnotes(mf) <- value

mf_nrheader(mf) <- value

mf_colgap(mf) <- value

mf_ncol(mf)

mf_nrow(mf)

mf_ncol(mf) <- value

\S4method{ncol}{MatrixPrintForm}(x)

mpf_has_rlabels(mf)

mf_has_rlabels(mf)
}
\arguments{
\item{mf}{(\code{MatrixPrintForm})\cr a \code{MatrixPrintForm} object.}

\item{value}{(\code{ANY})\cr the new value for the component in question.}

\item{x}{\code{MatrixPrintForm}. The object.}
}
\value{
\itemize{
\item Getters return the associated element of \code{mf}.
\item Setters return the modified \code{mf} object.
}
}
\description{
Most of these functions, particularly the setters, are intended almost exclusively for
internal use in, e.g., \code{\link{matrix_form}} methods, and should generally not be called by end users.
}
