% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{obj_round_type}
\alias{obj_round_type}
\alias{obj_round_type<-}
\title{Rounding Type}
\usage{
obj_round_type(obj)

obj_round_type(obj) <- value
}
\arguments{
\item{obj}{(\code{ANY})\cr a table-like object.}

\item{value}{The new rounding type of the object (see \code{\link[=round_fmt]{round_fmt()}} for details)}
}
\value{
The rounding type of the object (see \code{\link[=round_fmt]{round_fmt()}} for details).
}
\description{
When called on a table-like object using the formatters framework, this method returns the
rounding type of the object.
}
