% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{open_font_dev}
\alias{open_font_dev}
\alias{close_font_dev}
\title{Activate font state}
\usage{
open_font_dev(fontspec)

close_font_dev()
}
\arguments{
\item{fontspec}{font_spec. A font_spec object
specifying the font information to use for
calculating string widths and heights}
}
\value{
For \code{open_font_dev} a logical value indicating
whether a \emph{new} pdf device was opened; for \code{close_font_dev}, NULL. In both cases the value is returned invisibly.
}
\description{
Activate font state
}
\details{
The font device state is an environment with
four variables guaranteed to be set:

\describe{
\item{open}{logical(1). whether a device is already open with font info}
\item{fontspec}{font_spec. the font specification, if any, that is currently active (\code{list()} if none is).}
\item{spacewidth}{numeric(1). The width of the space character in the currently active font.}
\item{ismonospace}{logical(1). Whether the specified font is monospaced.}
}

\code{open_font_dev} opens a pdf device with the specified font
only if there is not one currently open with the same font.
If a new device is opened, it caches \code{spacewidth} and
\code{ismonospace} for use in \code{nchar_ttype}).

\code{close_font_dev} closes any open font state device
and clears the cached values.
}
\examples{
open_font_dev(font_spec("Times"))
nchar_ttype("Hiya there", font_spec("Times"))
close_font_dev()
}
