% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_size.R
\name{page_types}
\alias{page_types}
\title{List supported named page types}
\usage{
page_types()
}
\value{
a character vector of supported page types
}
\description{
List supported named page types
}
\examples{
page_types()
}
