% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats_xx.R
\name{round_sas}
\alias{round_sas}
\title{SAS rounding in R}
\usage{
round_sas(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{An integer specifying the number of decimal places to be
displayed after rounding. Default is 0.}
}
\value{
character vector of rounded values
}
\description{
Copied from tidytlg roundSAS to avoid importing
round_sas comes from this Stack Overflow post https://stackoverflow.com/questions/12688717/round-up-from-5
\cr Internal function only
}
\seealso{
\code{\link[=format_xx_sas]{format_xx_sas()}}, \code{\link[=format_sas_count_fraction]{format_sas_count_fraction()}}, \code{\link[=format_sas_count_denom_fraction]{format_sas_count_denom_fraction()}}
}
