% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_default_page_number}
\alias{set_default_page_number}
\title{Set default page number}
\usage{
set_default_page_number(page_number)
}
\arguments{
\item{page_number}{(\code{string})\cr Single string value to set the page number pattern}
}
\description{
Set default page number
}
