% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{spread_integer}
\alias{spread_integer}
\title{Spread an integer to a given length}
\usage{
spread_integer(x, len)
}
\arguments{
\item{x}{(\code{integer(1)})\cr number to spread.}

\item{len}{(\code{integer(1)})\cr number of times to repeat \code{x}.}
}
\value{
If \code{x} is a scalar whole number value (see \code{\link[=is.wholenumber]{is.wholenumber()}}), the value \code{x} is repeated \code{len} times.
Otherwise, an error is thrown.
}
\description{
Spread an integer to a given length
}
\examples{
spread_integer(3, 1)
spread_integer(0, 3)
spread_integer(1, 3)
spread_integer(2, 3)
spread_integer(3, 3)
spread_integer(4, 3)
spread_integer(5, 3)
spread_integer(6, 3)
spread_integer(7, 3)

}
