% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{sprintf_format}
\alias{sprintf_format}
\title{Specify text format via a \code{sprintf} format string}
\usage{
sprintf_format(format)
}
\arguments{
\item{format}{character(1). A format string passed to \code{sprintf}.}
}
\value{
A formatting function which wraps and will apply the specified \code{printf} style format
string \code{format}.
}
\description{
Specify text format via a \code{sprintf} format string
}
\examples{
fmtfun <- sprintf_format("(N=\%i")
format_value(100, format = fmtfun)

fmtfun2 <- sprintf_format("\%.4f - \%.2f")
format_value(list(12.23456, 2.724))

# You can also customise rtables values
fmtfun3 <- sprintf_format("this range => \%.1f")
format_value(100, format = fmtfun3)

fmtfun4 <- sprintf_format("sadbasd asa a \%.2f asda a \%.1f")
format_value(list(12.23456, 2.724), format = fmtfun4)
\dontrun{
require(rtable)
lyt <- basic_table() \%>\%
  split_cols_by("Species") \%>\%
  analyze("Sepal.Length", afun = function(x) {
    list(
      "mean (sd)" = rcell(c(mean(x), sd(x)), format = fmtfun4),
      "range" = rcell(diff(range(x)), format = fmtfun3)
    )
  })
build_table(lyt, iris)
}

}
\seealso{
\code{\link[base]{sprintf}}

\code{\link[formatters]{round_fmt}}

\code{\link[formatters]{list_formats}}
}
