% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_form.R
\name{test_matrix_form}
\alias{test_matrix_form}
\alias{basic_matrix_form}
\alias{basic_listing_mf}
\title{Create spoof matrix form from a data.frame}
\usage{
basic_matrix_form(
  df,
  parent_path = "root",
  ignore_rownames = FALSE,
  add_decoration = FALSE
)

basic_listing_mf(
  df,
  keycols = c("vs", "gear"),
  ignore_rownames = FALSE,
  add_decoration = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{parent_path}{character. parent path that all rows should be "children of",
defaults to \code{"root"}, and generally should not matter to end users.}

\item{ignore_rownames}{logical. If \code{TRUE}, rownames are ignored.}

\item{add_decoration}{logical. If \code{TRUE}, adds title and footers decorations to the
matrix form.}

\item{keycols}{character. Vector of \code{df} column names that are printed first and
repeated values are assigned to \code{""}. This format is characteristic of a listing matrix form.
When \code{NULL}, no key columns are used. Defaults to \code{c("vs", "gear")} for \code{mtcars} default dataset.}
}
\value{
A valid \code{MatrixPrintForm} object representing \code{df},
ready for ASCII rendering

A valid \code{MatrixPrintForm} object representing \code{df} as a listing,
ready for ASCII rendering.
}
\description{
Useful functions for writing testing/examples, and as a
starting point for more sophisticated custom \code{matrix_form} methods.
}
\section{Functions}{
\itemize{
\item \code{basic_listing_mf()}: Create a \code{MatrixPrintForm} object from a data.frame \code{df} that
respects the default formats for a listing object.

}}
\examples{
mform <- basic_matrix_form(mtcars)
cat(toString(mform))

mform <- basic_listing_mf(mtcars)
cat(toString(mform))

}
