% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{main_title}
\alias{main_title}
\alias{main_title,MatrixPrintForm-method}
\alias{main_title<-}
\alias{subtitles}
\alias{subtitles,MatrixPrintForm-method}
\alias{subtitles<-}
\alias{page_titles}
\alias{page_titles,MatrixPrintForm-method}
\alias{page_titles,ANY-method}
\alias{page_titles<-}
\alias{page_titles<-,MatrixPrintForm-method}
\alias{main_footer}
\alias{main_footer,MatrixPrintForm-method}
\alias{main_footer<-}
\alias{main_footer<-,MatrixPrintForm-method}
\alias{prov_footer}
\alias{prov_footer,MatrixPrintForm-method}
\alias{prov_footer<-}
\alias{prov_footer<-,MatrixPrintForm-method}
\alias{all_footers}
\alias{all_titles}
\title{General title/footer accessors}
\usage{
main_title(obj)

\S4method{main_title}{MatrixPrintForm}(obj)

main_title(obj) <- value

subtitles(obj)

\S4method{subtitles}{MatrixPrintForm}(obj)

subtitles(obj) <- value

page_titles(obj)

\S4method{page_titles}{MatrixPrintForm}(obj)

\S4method{page_titles}{ANY}(obj)

page_titles(obj) <- value

\S4method{page_titles}{MatrixPrintForm}(obj) <- value

main_footer(obj)

\S4method{main_footer}{MatrixPrintForm}(obj)

main_footer(obj) <- value

\S4method{main_footer}{MatrixPrintForm}(obj) <- value

prov_footer(obj)

\S4method{prov_footer}{MatrixPrintForm}(obj)

prov_footer(obj) <- value

\S4method{prov_footer}{MatrixPrintForm}(obj) <- value

all_footers(obj)

all_titles(obj)
}
\arguments{
\item{obj}{ANY. Object to extract information from.}

\item{value}{character. New value.}
}
\description{
General title/footer accessors
}
