% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{var_labels_remove}
\alias{var_labels_remove}
\title{Remove variable labels of a \code{data.frame}}
\usage{
var_labels_remove(x)
}
\arguments{
\item{x}{(\code{data.frame})\cr a \code{data.frame} object.}
}
\value{
\code{x} with its variable labels stripped.
}
\description{
Remove \code{label} attribute from all variables in a data frame.
}
\examples{
x <- var_labels_remove(iris)

}
