% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{var_relabel}
\alias{var_relabel}
\title{Copy and change variable labels of a \code{data.frame}}
\usage{
var_relabel(x, ...)
}
\arguments{
\item{x}{(\code{data.frame})\cr a data frame object.}

\item{...}{name-value pairs, where each name corresponds to a variable name in
\code{x} and the value to the new variable label.}
}
\value{
A copy of \code{x} with labels modified according to \code{...}
}
\description{
Relabel a subset of the variables.
}
\examples{
x <- var_relabel(iris, Sepal.Length = "Sepal Length of iris flower")
var_labels(x)

}
