% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{with_label}
\alias{with_label}
\title{Return an object with a label attribute}
\usage{
with_label(x, label)
}
\arguments{
\item{x}{(\code{ANY})\cr an object.}

\item{label}{(\code{string})\cr label attribute to attach to \code{x}.}
}
\value{
\code{x} labeled by \code{label}. Note that the exact mechanism of labeling should be considered
an internal implementation detail, but the label can always be retrieved via \code{obj_label}.
}
\description{
Return an object with a label attribute
}
\examples{
x <- with_label(c(1, 2, 3), label = "Test")
obj_label(x)

}
