% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{split_word_ttype}
\alias{split_word_ttype}
\alias{wrap_string_ttype}
\title{wrap string given a Truetype font}
\usage{
split_word_ttype(string, width_spc, fontspec, min_ok_chars)

wrap_string_ttype(
  string,
  width_spc,
  fontspec,
  collapse = NULL,
  min_ok_chars = min(floor(nchar(string)/2), 4, floor(width_spc/2))
)
}
\arguments{
\item{fontspec}{font_spec. A font_spec object
specifying the font information to use for
calculating string widths and heights, as
returned by \code{\link[=font_spec]{font_spec()}}}

\item{min_ok_chars}{numeric(1). Number of minimum characters which much remain
on either side when a word is split}

\item{collapse}{character(1) or \code{NULL}. If the words that have been split should
be pasted together with the collapse character. This is usually done internally
with \code{"\\n"} to have the wrapping updated along with other internal values.}
}
\value{
\code{string}, broken up into a word-wrapped vector
}
\description{
wrap string given a Truetype font
}
