% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_form.R
\name{basic_matrix_form}
\alias{basic_matrix_form}
\title{Create spoof matrix form from a data.frame}
\usage{
basic_matrix_form(
  df,
  parent_path = "root",
  fontspec = font_spec(),
  rowlabels = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{parent_path}{character. parent path that all rows should be "children of",
defaults to \code{"root"}, and generally should not matter to end users.}

\item{fontspec}{font_spec. A font_spec object
specifying the font information to use for
calculating string widths and heights, as
returned by \code{\link[=font_spec]{font_spec()}}}

\item{rowlabels}{logical(1). Should the row labels of \code{df} be used as a
row-label pseudo column? Defaults to \code{TRUE}}
}
\value{
A valid \code{MatrixPrintForm} object representing \code{df},
ready for ASCII rendering
}
\description{
This is  useful primarily  for writing  testing/examples, and as a
starting point for more sophisticated custom \code{matrix_form} methods
}
\examples{
mform <- basic_matrix_form(mtcars)
cat(toString(mform))
}
