% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{decimal_align}
\alias{decimal_align}
\title{Decimal Alignment}
\usage{
decimal_align(string_mat, align_mat)
}
\arguments{
\item{string_mat}{string matrix component of matrix print form object.}

\item{align_mat}{aligns matrix component of matrix print form object.
Should contain either \code{dec_left}, \code{dec_right} or \code{decimal} for values to be decimal aligned.}
}
\value{
Processed string matrix of matrix print form with decimal aligned values.
}
\description{
Aligning decimal values of string matrix. Allowed alignments are: \code{dec_left},
\code{dec_right} and \code{decimal}.
}
\examples{

dfmf <- basic_matrix_form(mtcars)
dfmf$aligns[, -c(1)] <- "dec_left"
decimal_align(dfmf$strings, dfmf$aligns)

}
