% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{default_page_number}
\alias{default_page_number}
\alias{set_default_page_number}
\title{Default page number format}
\usage{
default_page_number()

set_default_page_number(page_number)
}
\arguments{
\item{page_number}{character(1). Single string value to set the page number format.
It should be like the following format: \code{"page {i}/{n}"}. \code{{i}} will be replaced with
current page number, \code{{n}} will be replaced with total page number. It uses current \code{cpp}
to position the string on the bottom right corner.}
}
\value{
The page number format string. If not set, it returns \code{NULL}.
}
\description{
If set the default page number string will appear on the bottom right of
every paginated table. It uses the current \code{cpp} to position the string.
}
\examples{
default_page_number()
set_default_page_number("page {i} of {n}")
default_page_number()

}
