% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{default_page_number}
\alias{default_page_number}
\alias{set_default_page_number}
\title{Default page number format}
\usage{
default_page_number()

set_default_page_number(page_number)
}
\arguments{
\item{page_number}{(\code{string})\cr single string value to set the page number format.
It should be formatted similarly to the following format: \code{"page {i}/{n}"}.
\code{{i}} will be replaced with the current page number, and \code{{n}} will be replaced with the
total page number. Current \code{cpp} is used to position the string in the bottom right corner.}
}
\value{
The page number format string (\code{NULL} if not set).
}
\description{
If set, the default page number string will appear on the bottom right of
every page of a paginated table. The current \code{cpp} is used to position the string.
}
\examples{
default_page_number()
set_default_page_number("page {i} of {n}")
default_page_number()

}
