% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{divider_height}
\alias{divider_height}
\alias{divider_height,ANY-method}
\title{Divider height}
\usage{
divider_height(obj)

\S4method{divider_height}{ANY}(obj)
}
\arguments{
\item{obj}{(\code{ANY})\cr object.}
}
\value{
The height, in lines of text, of the divider between header and body. Currently
returns \code{1L} for the default method.
}
\description{
Divider height
}
\examples{
divider_height(mtcars)

}
