% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exporters.R
\name{export_as_rtf}
\alias{export_as_rtf}
\title{Export table to RTF}
\usage{
export_as_rtf(
  tt,
  file = NULL,
  colwidths = propose_column_widths(matrix_form(tt, TRUE)),
  page_type = "letter",
  pg_width = page_dim(page_type)[if (landscape) 2 else 1],
  pg_height = page_dim(page_type)[if (landscape) 1 else 2],
  landscape = FALSE,
  margins = c(bottom = 4, left = 4, top = 4, right = 4),
  font_size = 8,
  font_family = "Courier",
  ...
)
}
\arguments{
\item{tt}{TableTree (or related class). A TableTree object representing a populated table.}

\item{file}{character(1). File to write.}

\item{colwidths}{numeric vector. Column widths for use with vertical pagination. Currently ignored.}

\item{margins}{ A numeric vector interpreted in the same way
    as \code{par(mar)} in base graphics. }

\item{...}{Passed directly to \code{\link{paginate_table}}}
}
\description{
Experimental export to the RTF format.
}
\details{
RTF export occurs by via the following steps

\itemize{
\item{the table is paginated to the page size (Vertically and horizontally)}
\item{Each separate page is converted to a MatrixPrintForm and from there to RTF-encoded text}
\item{Separate rtfs text chunks are combined and written out as a single RTF file}
}

Conversion of \code{MatrixPrintForm} objects to RTF is done via \code{\link[=mpf_to_rtf]{mpf_to_rtf()}}.
}
