% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{fmt_config}
\alias{fmt_config}
\title{Format configuration}
\usage{
fmt_config(format = NULL, na_str = "NA", align = "center")
}
\arguments{
\item{format}{(\code{string} or \code{function})\cr a format label (string) or formatter function.}

\item{na_str}{(\code{string})\cr string that should be displayed in place of missing values.}

\item{align}{(\code{string})\cr alignment values should be rendered with.}
}
\value{
An object of class \code{fmt_config} which contains the following elements:
\itemize{
\item \code{format}
\item \code{na_str}
\item \code{align}
}
}
\description{
Format configuration
}
\examples{
fmt_config(format = "xx.xx", na_str = "-", align = "left")
fmt_config(format = "xx.xx - xx.xx", align = "right")

}
