% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{formatters-package}
\alias{formatters}
\alias{formatters-package}
\title{formatters Package}
\description{
Package to format tables and listings in a flexible way.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/formatters/}
  \item \url{https://github.com/insightsengineering/formatters/}
  \item Report bugs at \url{https://github.com/insightsengineering/formatters/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Gabriel Becker \email{gabembecker@gmail.com} (original creator of the package)
  \item Adrian Waddell \email{adrian.waddell@gene.com}
  \item Davide Garolini \email{davide.garolini@roche.com} (\href{https://orcid.org/0000-0002-1445-1369}{ORCID})
  \item Emily de la Rua \email{emily.de_la_rua@contractors.roche.com} (\href{https://orcid.org/0009-0000-8738-5561}{ORCID})
}

Other contributors:
\itemize{
  \item Abinaya Yogasekaram \email{abinaya.yogasekaram@contractors.roche.com} (\href{https://orcid.org/0009-0005-2083-1105}{ORCID}) [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
