% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ifnotlen0}
\alias{ifnotlen0}
\alias{\%||\%}
\title{\verb{\%||\%} (if length-0) alternative operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{(\code{ANY})\cr element to select \emph{only} if it is not of length 0.}

\item{b}{(\code{ANY})\cr element to select if \code{a} has length 0.}
}
\value{
\code{a} if it is not of length 0, otherwise \code{b}.
}
\description{
\verb{\%||\%} (if length-0) alternative operator
}
\examples{
6 \%||\% 10

character() \%||\% "hi"

NULL \%||\% "hi"

}
