% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{\code{is.wholenumber}}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{numeric(1). A numeric value}

\item{tol}{numeric(1). A precision tolerance.}
}
\value{
\code{TRUE} if \code{x} is within \code{tol} of zero,
\code{FALSE} otherwise.
}
\description{
\code{is.wholenumber}
}
\examples{
is.wholenumber(5)
is.wholenumber(5.00000000000000001)
is.wholenumber(.5)

}
