% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lab_name}
\alias{lab_name}
\alias{obj_name}
\alias{obj_name<-}
\alias{obj_label}
\alias{obj_label<-}
\alias{obj_label,ANY-method}
\alias{obj_label<-,ANY-method}
\alias{obj_format}
\alias{obj_format,ANY-method}
\alias{obj_format,fmt_config-method}
\alias{obj_format<-}
\alias{obj_format<-,ANY-method}
\alias{obj_format<-,fmt_config-method}
\alias{obj_na_str}
\alias{obj_na_str,ANY-method}
\alias{obj_na_str,fmt_config-method}
\alias{obj_na_str<-}
\alias{obj_na_str<-,ANY-method}
\alias{obj_na_str<-,fmt_config-method}
\alias{obj_align}
\alias{obj_align,ANY-method}
\alias{obj_align,fmt_config-method}
\alias{obj_align<-}
\alias{obj_align<-,ANY-method}
\alias{obj_align<-,fmt_config-method}
\title{Label, name, and format accessor generics}
\usage{
obj_name(obj)

obj_name(obj) <- value

obj_label(obj)

obj_label(obj) <- value

\S4method{obj_label}{ANY}(obj)

\S4method{obj_label}{ANY}(obj) <- value

obj_format(obj)

\S4method{obj_format}{ANY}(obj)

\S4method{obj_format}{fmt_config}(obj)

obj_format(obj) <- value

\S4method{obj_format}{ANY}(obj) <- value

\S4method{obj_format}{fmt_config}(obj) <- value

obj_na_str(obj)

\S4method{obj_na_str}{ANY}(obj)

\S4method{obj_na_str}{fmt_config}(obj)

obj_na_str(obj) <- value

\S4method{obj_na_str}{ANY}(obj) <- value

\S4method{obj_na_str}{fmt_config}(obj) <- value

obj_align(obj)

\S4method{obj_align}{ANY}(obj)

\S4method{obj_align}{fmt_config}(obj)

obj_align(obj) <- value

\S4method{obj_align}{ANY}(obj) <- value

\S4method{obj_align}{fmt_config}(obj) <- value
}
\arguments{
\item{obj}{(\code{ANY})\cr the object.}

\item{value}{character(1). The new label}
}
\value{
The name, format, or label of \code{obj} for getters, or \code{obj} after modification for setters.
}
\description{
Getters and setters for basic, relatively universal attributes of "table-like" objects.
}
\seealso{
with_label
}
