% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_form.R
\name{mf_strings}
\alias{mf_strings}
\alias{mf_spans}
\alias{mf_aligns}
\alias{mf_display}
\alias{mf_formats}
\alias{mf_rinfo}
\alias{mf_has_topleft}
\alias{mf_lgrouping}
\alias{mf_rfnotes}
\alias{mf_nlheader}
\alias{mf_nrheader}
\alias{mf_strings<-}
\alias{mf_spans<-}
\alias{mf_aligns<-}
\alias{mf_display<-}
\alias{mf_formats<-}
\alias{mf_rinfo<-}
\alias{mf_lgrouping<-}
\alias{mf_rfnotes<-}
\alias{mf_nrheader<-}
\alias{mpf_has_rlabels}
\title{Setters and Getters for aspects of MatrixPrintForm Objects}
\usage{
mf_strings(mf)

mf_spans(mf)

mf_aligns(mf)

mf_display(mf)

mf_formats(mf)

mf_rinfo(mf)

mf_has_topleft(mf)

mf_lgrouping(mf)

mf_rfnotes(mf)

mf_nlheader(mf)

mf_nrheader(mf)

mf_strings(mf) <- value

mf_spans(mf) <- value

mf_aligns(mf) <- value

mf_display(mf) <- value

mf_formats(mf) <- value

mf_rinfo(mf) <- value

mf_lgrouping(mf) <- value

mf_rfnotes(mf) <- value

mf_nrheader(mf) <- value

mpf_has_rlabels(mf)
}
\arguments{
\item{mf}{MatrixPrintForm(1). A MatrixPrintForm object}

\item{value}{ANY. The new value for the component in question.}
}
\value{
The element of the MatrixPrintForm associated with the getter, or
the modified MatrixPrintForm object in the case of a setter.
}
\description{
Most of these functions, particularly the settters, are intended
almost exclusively for internal use in, e.g., `matrix_form` methods,
and should generally not be called by end users.
}
