% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nlines}
\alias{nlines}
\alias{nlines,list-method}
\alias{nlines,NULL-method}
\alias{nlines,character-method}
\title{Number of lines required to print a value}
\usage{
nlines(x, colwidths = NULL, max_width = NULL, fontspec, col_gap)

\S4method{nlines}{list}(x, colwidths = NULL, max_width = NULL, fontspec, col_gap = NULL)

\S4method{nlines}{NULL}(x, colwidths = NULL, max_width = NULL, fontspec, col_gap = NULL)

\S4method{nlines}{character}(x, colwidths = NULL, max_width = NULL, fontspec, col_gap = NULL)
}
\arguments{
\item{x}{ANY. The object to be printed}

\item{colwidths}{numeric. Column widths (if necessary).}

\item{max_width}{numeric(1). Width strings should be wrapped to
when determining how many lines they require.}

\item{fontspec}{font_spec. A font_spec object
specifying the font information to use for
calculating string widths and heights, as
returned by \code{\link[=font_spec]{font_spec()}}}

\item{col_gap}{numeric(1). Width of gap between columns in number of spaces.
Only used by methods which must calculate span widths after wrapping.}
}
\value{
A scalar numeric indicating the number of lines needed
to render the object \code{x}.
}
\description{
Number of lines required to print a value
}
