% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nlines}
\alias{nlines}
\alias{nlines,list-method}
\alias{nlines,NULL-method}
\alias{nlines,character-method}
\title{Number of lines required to print a value}
\usage{
nlines(x, colwidths)

\S4method{nlines}{list}(x, colwidths)

\S4method{nlines}{`NULL`}(x, colwidths)

\S4method{nlines}{character}(x, colwidths)
}
\arguments{
\item{x}{ANY. The object to be printed}

\item{colwidths}{numeric. Column widths (if necessary)}
}
\value{
A scalar numeric indicating the number of lines needed
to render the object \code{x}.
}
\description{
Number of lines required to print a value
}
