% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nlines}
\alias{nlines}
\alias{nlines,list-method}
\alias{nlines,NULL-method}
\alias{nlines,character-method}
\title{Number of lines required to print a value}
\usage{
nlines(x, colwidths = NULL, max_width = NULL)

\S4method{nlines}{list}(x, colwidths = NULL, max_width = NULL)

\S4method{nlines}{NULL}(x, colwidths = NULL, max_width = NULL)

\S4method{nlines}{character}(x, colwidths = NULL, max_width = NULL)
}
\arguments{
\item{x}{(\code{any})\cr the object to be printed.}

\item{colwidths}{(\code{numeric})\cr column widths (if necessary).}

\item{max_width}{(\code{numeric(1)})\cr width that strings should be wrapped to when
determining how many lines they require.}
}
\value{
The number of lines needed to render the object \code{x}.
}
\description{
Number of lines required to print a value
}
