% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_size.R
\name{page_lcpp}
\alias{page_lcpp}
\title{Determine LPP and CPP based on font and page type}
\usage{
page_lcpp(
  page_type = page_types(),
  landscape = FALSE,
  font_family = "Courier",
  font_size = 12,
  lineheight = 1,
  margins = c(top = 0.5, bottom = 0.5, left = 0.75, right = 0.75),
  pg_width = NULL,
  pg_height = NULL
)
}
\arguments{
\item{page_type}{character(1).   Name   of  a  page  type.   See
`page_types`.   Ignored when  `pg_width` and  `pg_height`
are set directly.}

\item{landscape}{logical(1). Should the  dimensions of `page_type`
be inverted  for landscape?  Defaults to  `FALSE`, ignored when
`pg_width` and `pg_height` are set directly.}

\item{font_family}{character(1). Name of a font family. An error
will be thrown if the family named is not monospaced. Defaults
to Courier.}

\item{font_size}{numeric(1). Fontsize, defaults to 12.}

\item{lineheight}{numeric(1). Lineheight, defaults to 1.}

\item{margins}{numeric(4). Named numeric vector containing `'top'`,
`'bottom'`, `'left'`, and `'right'` margins in inches. Defaults
to `.5` inches for both vertical margins and `.75` for both
horizontal margins.}

\item{pg_width}{numeric(1). Page width in inches.}

\item{pg_height}{numeric(1). Page height in inches.}
}
\value{
a named list containing `lpp` and `cpp` elements suitable
for use by the pagination machinery.
}
\description{
Determine LPP and CPP based on font and page type
}
\examples{
page_lcpp()
page_lcpp(font_size = 10)
page_lcpp("a4", font_size = 10)

page_lcpp(margins = c(top = 1, bottom = 1, left = 1, right = 1))
page_lcpp(pg_width = 10, pg_height = 15)

}
