% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_size.R
\name{page_types}
\alias{page_types}
\alias{page_dim}
\title{Supported named page types}
\usage{
page_types()

page_dim(page_type)
}
\arguments{
\item{page_type}{(\code{string})\cr the name of a page size specification. Call
\code{\link[=page_types]{page_types()}} for supported values.}
}
\value{
\itemize{
\item \code{page_types} returns a character vector of supported page types
\item \code{page_dim} returns the dimensions (width, then height) of the selected page type.
}
}
\description{
List supported named page types.
}
\examples{
page_types()
page_dim("a4")
}
