% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{paginate_indices}
\alias{paginate_indices}
\alias{paginate}
\alias{pagination}
\alias{paginate_to_mpfs}
\title{Paginate a table-like object for rendering}
\usage{
paginate_indices(
  obj,
  page_type = "letter",
  font_family = "Courier",
  font_size = 8,
  lineheight = 1,
  landscape = FALSE,
  pg_width = NULL,
  pg_height = NULL,
  margins = c(top = 0.5, bottom = 0.5, left = 0.75, right = 0.75),
  lpp = NULL,
  cpp = NULL,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  tf_wrap = FALSE,
  max_width = NULL,
  indent_size = 2,
  pg_size_spec = NULL,
  rep_cols = num_rep_cols(obj),
  col_gap = 2,
  verbose = FALSE
)

paginate_to_mpfs(
  obj,
  page_type = "letter",
  font_family = "Courier",
  font_size = 8,
  lineheight = 1,
  landscape = FALSE,
  pg_width = NULL,
  pg_height = NULL,
  margins = c(top = 0.5, bottom = 0.5, left = 0.75, right = 0.75),
  lpp = NULL,
  cpp = NULL,
  min_siblings = 2,
  nosplitin = character(),
  colwidths = NULL,
  tf_wrap = FALSE,
  max_width = NULL,
  indent_size = 2,
  pg_size_spec = NULL,
  rep_cols = num_rep_cols(obj),
  col_gap = 2,
  verbose = FALSE
)
}
\arguments{
\item{obj}{ANY.  object   to   be  paginated.   Must  have   a
\code{\link{matrix_form}} method.}

\item{page_type}{character(1).   Name   of  a  page  type.   See
\code{page_types}.   Ignored when  \code{pg_width} and  \code{pg_height}
are set directly.}

\item{font_family}{character(1). Name of a font family. An error
will be thrown if the family named is not monospaced. Defaults
to Courier.}

\item{font_size}{numeric(1). Font size, defaults to 12.}

\item{lineheight}{numeric(1). Line height, defaults to 1.}

\item{landscape}{logical(1). Should the  dimensions of \code{page_type}
be inverted  for landscape?  Defaults to  \code{FALSE}, ignored when
\code{pg_width} and \code{pg_height} are set directly.}

\item{pg_width}{numeric(1). Page width in inches.}

\item{pg_height}{numeric(1). Page height in inches.}

\item{margins}{numeric(4). Named numeric vector containing \code{'bottom'},
\code{'left'}, \code{'top'}, and \code{'right'} margins in inches. Defaults
to \code{.5} inches for both vertical margins and \code{.75} for both
horizontal margins.}

\item{lpp}{numeric(1) or NULL. Lines per page. if NULL (the default,
this is calculated automatically based on the specified page
size).}

\item{cpp}{numeric(1). Number of characters per page (width)}

\item{min_siblings}{numeric. Minimum sibling rows which must appear on either side of pagination row for a
mid-subtable split to be valid. Defaults to 2.}

\item{nosplitin}{character. List of names of sub-tables where page-breaks are not allowed, regardless of other
considerations. Defaults to none.}

\item{colwidths}{numeric vector.  Column widths (in characters) for
use with vertical pagination.}

\item{tf_wrap}{logical(1). Should  the texts for  title, subtitle,
and footnotes be wrapped?}

\item{max_width}{integer(1), character(1) or NULL. Width that title
and   footer   (including   footnotes)  materials   should   be
word-wrapped to. If NULL, it is  set to the current print width
of the  session (\code{getOption("width")}). If set to \code{"auto"},
the width of the table (plus any table inset) is used. Ignored
completely if \code{tf_wrap} is \code{FALSE}.}

\item{indent_size}{numeric(1). Indent size in characters. Ignored
when \code{x} is already a MatrixPrintForm object in favor of information
there.}

\item{pg_size_spec}{page_size_spec. A pre-calculated page
size specification. Typically this is not set in end user code.}

\item{rep_cols}{numeric(1). Number of \emph{columns} (not including
row labels) to be repeated on every page. Defaults to 0}

\item{col_gap}{numeric(1). Space (in characters) between columns}

\item{verbose}{logical(1). Should additional informative messages about the search for
pagination breaks be shown. Defaults to \code{FALSE}.}
}
\value{
for \code{paginate_indices} a list with two elements of the same
length:   \code{pag_row_indices},    and   \code{pag_col_indices}.    For
\code{paginate_to_mpfs},   a  list   of  \code{MatrixPrintForm} objects
representing each individual page after pagination (including
forced pagination if necessary).
}
\description{
This functions prepares \code{obj} for paginated rendering,
by paginating it \code{emph} into multiple \code{MatrixPrintForm}
objects representing individual pages.
}
\details{
\code{paginate_to_mpfs} supports objects which require 'forced pagination', ie
pagination at points not dictated by page dimensions. It does this by
automatically calling the \code{do_force_pagination} generic.

\code{paginate_indices},  on  the  other hand,  \emph{does  not  support
forced pagination},  because it returns  only a set of  indices for
row and column subsetting for each page, and thus cannot retain any
changes, e.g., to titles, done within \code{do_forced_paginate}.
}
\examples{
mpf <- basic_matrix_form(mtcars)

paginate_indices(mpf, pg_width = 5, pg_height = 3)

paginate_to_mpfs(mpf, pg_width = 5, pg_height = 3)
}
