% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{propose_column_widths}
\alias{propose_column_widths}
\title{Propose Column Widths based on an object's `MatrixPrintForm` form}
\usage{
propose_column_widths(x, indent_size = 2)
}
\arguments{
\item{x}{MatrixPrintForm object, or an object with a `matrix_form`
method.}

\item{indent_size}{numeric(1). Indent size in characters. Ingored
when `x` is already a MatrixPrintForm object in favor of information
there.}
}
\value{
a vector of column widths based on the content of \code{x}
for use in printing and pagination.
}
\description{
The row names are also considered a column for the output
}
\examples{
mf <- basic_matrix_form(mtcars)
propose_column_widths(mf)

# ' @examples
# ' library(dplyr)
# ' library(rtables)
# ' iris2 <- iris \%>\%
# '   group_by(Species) \%>\%
# '   mutate(group = as.factor(rep_len(c("a", "b"), length.out = n()))) \%>\%
# '   ungroup()
# '
# ' l <- basic_table() \%>\%
# '   split_cols_by("Species") \%>\%
# '   split_cols_by("group") \%>\%
# '   analyze(c("Sepal.Length", "Petal.Width"), afun = list_wrap_x(summary) , format = "xx.xx")
# '
# ' tbl <- build_table(l, iris2)
# ' mf <- matrix_form(tbl)
# ' propose_column_widths(mf)
}
