% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{propose_column_widths}
\alias{propose_column_widths}
\title{Propose Column Widths based on an object's \code{MatrixPrintForm} form}
\usage{
propose_column_widths(x, indent_size = 2)
}
\arguments{
\item{x}{MatrixPrintForm object, or an object with a \code{matrix_form}
method.}

\item{indent_size}{numeric(1). Indent size in characters. Ingored
when \code{x} is already a MatrixPrintForm object in favor of information
there.}
}
\value{
a vector of column widths based on the content of \code{x}
for use in printing and pagination.
}
\description{
The row names are also considered a column for the output
}
\examples{
mf <- basic_matrix_form(mtcars)
propose_column_widths(mf)

}
