% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tostring.R
\name{spread_integer}
\alias{spread_integer}
\title{spread \code{x} into \code{len} elements}
\usage{
spread_integer(x, len)
}
\arguments{
\item{x}{\code{numeric(1)}. The number to spread}

\item{len}{\code{numeric(1)}. The number of times to repeat \code{x}}
}
\value{
if \code{x} is a scalar "whole number" value (see \code{\link{is.wholenumber}}),
the value \code{x} repeated \code{len} times. If not, an error is thrown.
}
\description{
spread \code{x} into \code{len} elements
}
\examples{
spread_integer(3, 1)
spread_integer(0, 3)
spread_integer(1, 3)
spread_integer(2, 3)
spread_integer(3, 3)
spread_integer(4, 3)
spread_integer(5, 3)
spread_integer(6, 3)
spread_integer(7, 3)
}
