% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{sprintf_format}
\alias{sprintf_format}
\title{Specify text format via a \code{sprintf} format string}
\usage{
sprintf_format(format)
}
\arguments{
\item{format}{(\code{character(1)})\cr a format string passed to \code{\link[=sprintf]{sprintf()}}.}
}
\value{
A formatting function which wraps and will apply the specified \code{sprintf}-style format
to string \code{format}.
}
\description{
Specify text format via a \code{sprintf} format string
}
\examples{
fmtfun <- sprintf_format("(N=\%i")
format_value(100, format = fmtfun)

fmtfun2 <- sprintf_format("\%.4f - \%.2f")
format_value(list(12.23456, 2.724))

}
\seealso{
\code{\link[=sprintf]{sprintf()}}
}
