% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_form.R
\name{test_matrix_form}
\alias{test_matrix_form}
\alias{basic_matrix_form}
\alias{basic_listing_mf}
\title{Create spoof matrix form from a data frame}
\usage{
basic_matrix_form(
  df,
  parent_path = "root",
  ignore_rownames = FALSE,
  add_decoration = FALSE
)

basic_listing_mf(df, keycols = names(df)[1], add_decoration = TRUE)
}
\arguments{
\item{df}{(\code{data.frame})\cr a data frame.}

\item{parent_path}{(\code{string})\cr parent path that all rows should be "children of".
Defaults to \code{"root"}, and generally should not matter to end users.}

\item{ignore_rownames}{(\code{flag})\cr whether rownames should be ignored.}

\item{add_decoration}{(\code{flag})\cr whether adds title and footer decorations should
be added to the matrix form.}

\item{keycols}{(\code{character})\cr a vector of \code{df} column names that are printed first and for which
repeated values are assigned \code{""}. This format is characteristic of a listing matrix form.}
}
\value{
A valid \code{MatrixPrintForm} object representing \code{df} that is ready for ASCII rendering.

A valid \code{MatrixPrintForm} object representing \code{df} as a listing that is ready for ASCII
rendering.
}
\description{
Useful functions for writing tests and examples, and a starting point for
more sophisticated custom \code{matrix_form} methods.
}
\section{Functions}{
\itemize{
\item \code{basic_listing_mf()}: Create a \code{MatrixPrintForm} object from data frame \code{df} that
respects the default formats for a listing object.

}}
\examples{
mform <- basic_matrix_form(mtcars)
cat(toString(mform))

mform <- basic_listing_mf(mtcars)
cat(toString(mform))

}
