% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/tostring.R
\name{toString}
\alias{toString}
\alias{toString,MatrixPrintForm-method}
\title{toString}
\usage{
toString(x, ...)

\S4method{toString}{MatrixPrintForm}(
  x,
  widths = NULL,
  wrap = FALSE,
  max_width = NULL,
  col_gap = x$col_gap,
  hsep = default_hsep()
)
}
\arguments{
\item{x}{ANY. Object to be prepared for rendering.}

\item{...}{Passed to individual methods.}

\item{widths}{numeric (or NULL). (proposed) widths for the columns of
\code{x}. The expected length of this numeric vector can be retrieved with
`ncol() + 1` as the column of row names must also be considered.}

\item{wrap}{logical(1). Should the texts like title and footnotes be wrapped?}

\item{max_width}{integer(1) (or NULL). If NULL, it will be assigned to
`options("width")`. Otherwise, it is also possible to use `c("auto")` to
automatically use the table width.}

\item{col_gap}{numeric(1). Space (in characters) between columns}

\item{hsep}{character(1). Characters to repeat to create header/body
separator line.}
}
\value{
A character string containing the ASCII rendering
of the table-like object reprseented by `x`
}
\description{
Transform a complex object into a string representation ready
to be printed or written to a plain-text file
}
\examples{
mform <- basic_matrix_form(mtcars)
cat(toString(mform))
}
