% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagination.R
\name{vert_pag_indices}
\alias{vert_pag_indices}
\title{Find Column Indices for Vertical Pagination}
\usage{
vert_pag_indices(
  obj,
  cpp = 40,
  colwidths = NULL,
  verbose = FALSE,
  rep_cols = 0L,
  fontspec,
  nosplitin = character()
)
}
\arguments{
\item{obj}{ANY.  object   to   be  paginated.   Must  have   a
\code{\link{matrix_form}} method.}

\item{cpp}{numeric(1). Number of characters per page (width)}

\item{colwidths}{numeric vector.  Column widths (in characters) for
use with vertical pagination.}

\item{verbose}{logical(1). Should additional informative messages about the search for
pagination breaks be shown. Defaults to \code{FALSE}.}

\item{rep_cols}{numeric(1). Number of \emph{columns} (not including
row labels) to be repeated on every page. Defaults to 0}

\item{nosplitin}{character. List of names of sub-tables where page-breaks are not allowed, regardless of other
considerations. Defaults to none.}
}
\value{
A list partitioning the vector of column indices
into subsets for 1 or more horizontally paginated pages.
}
\description{
Find Column Indices for Vertical Pagination
}
\examples{
mf <- basic_matrix_form(df = mtcars)
colpaginds <- vert_pag_indices(mf, fontspec = font_spec())
lapply(colpaginds, function(j) mtcars[, j, drop = FALSE])
}
